requirements.txt
django-1.5.4
allauth
django-dajaxice
django-dajax
